Criando classe  FightAttributes, que serão os atributos que serão modificados durante a luta

+++ MMASimulation.Shared/Models/Fights/FightPBP.cs
@@ -3,8 +3,8 @@
     public class FightPBP
     {
 
-        public long Id { get; set; }
-        public long FightId { get; set; }
+        public int Id { get; set; }
+        public int FightId { get; set; }
         public required Fight Fight { get; set; }
         public string PbpData { get; set; } = string.Empty;
